Updated types to swell-js and bump version
