+++ types/swell-js/locale.d.ts
@@ -1,0 +1,6 @@
+export interface Locale {
+    cartId?: string;
+    lastCheckoutId?: string;
+    locale: string;
+    publicKey: string;
+}
